feat: fix hooks imports, rename MemoryBoard, cleanup utils
